New NMI Editor: Add and Delete of Controls to GroupThing in Preview Support for update of Fields with different Owner Thing in Forms new OPC UA Thing - moved all OPC UA releated classes to new cdeUAThing.cs Preview of TheThingGroup: HMI Screen Support within the NMI
